app/requirements.txt
Django==2.2.8
djangorestframework==3.8.2
Pillow==6.1.0
psycopg2